requirements.txt
grpcio-tools
beanie
pydantic